% Generated by roxygen2 (4.0.0): do not edit by hand

\name{blank_theme}
\alias{blank_theme}
\title{Return a blank ggplot theme, good for overlaying maps on top of.}
\usage{
blank_theme(bgfill = "white")
}
\arguments{
  \item{bgfill}{Background fill color.}
}
\description{
Return a blank ggplot theme, good for overlaying maps on
top of.
}

