% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get_ggbasemap}
\alias{get_ggbasemap}
\title{Get basemap using the OpenStreetMap package. Also cache it into a filename unless
cache = FALSE, downloading+projecting tiles takes a while so caching generally makes sense.}
\usage{
get_ggbasemap(lat_range, lon_range, num_tiles = 9, type = "mapbox",
  cache = TRUE, verbose = FALSE)
}
\arguments{
\item{lat_range}{A vector of minimum and maximum latitude to download basemap for.}

\item{lat_range}{A vector of minimum and maximum latitude to download basemap for.}

\item{type}{What kind of basemap? See options in OpenStreetMap package. "none" = blank.}

\item{num_tiles}{How many tiles (at least) to download?}

\item{cache}{Whether or not to cache the downloaded map. TRUE by default. Saves
an .RDS file that encodes latrange/lonrange.}

\item{verbose}{Prints message about what is being done if TRUE. FALSE by default.}
}
\value{
A ggplot object.
}
\description{
Get basemap using the OpenStreetMap package. Also cache it into a filename unless
cache = FALSE, downloading+projecting tiles takes a while so caching generally makes sense.
}

