% Generated by roxygen2 (4.0.0): do not edit by hand
\name{read_OSM}
\alias{read_OSM}
\title{Create a data.table with columns as specified, given osm file.}
\usage{
read_OSM(osm_file, osmconvert = "osmconvert",
  columns = "@lat @lon @timestamp")
}
\arguments{
\item{osm_file}{The osm_file to process. Can be a .csv, .osm, or .pbf file.}

\item{osmconvert}{Path to the osmconvert command-line utility. See
http://wiki.openstreetmap.org/wiki/Osmconvert for installation.}

\item{columns}{Columns that osmconvert should output. Syntax from osmconvert.}
}
\value{
A data.table, with lat, lon, and timestamp columns.
}
\description{
OSM file can be a .csv file, a .osm file or a .pbf file.
If .osm or .pbf file are input, then osmconvert must be the path to
the osmconvert command. A .csv file with an identical basename will be created.
}

