% Generated by roxygen2 (4.0.0): do not edit by hand
\name{time_lapse}
\alias{time_lapse}
\title{Function to create a time lapse animation out of a data table of nodes.}
\usage{
time_lapse(node_data_table, time_unit = "year", basemap_type = "mapbox",
  plot_single_timeunit_FUN = plot_single_timeunit, verbose = FALSE, ...)
}
\arguments{
\item{node_data_table}{data.table containing, at least lat, lon, and timestamp. lat + lon must
be in WGS84. timestamp must be of type R POSIXct.}

\item{time_unit}{Time unit for each frame. As per lubridate::round_date, should be one of
"second","minute","hour","day", "week", "month", or "year."}

\item{plot_single_timeunit_FUN}{Function for plotting a single frame. See plot_single_timeunit for
details on what the function should look like.}

\item{basemap_type}{Which OSM basemap to download? type must be supported by OpenStreetMap
package. Default is mapbox. Pass "none" to skip basemap.}

\item{verbose}{A flag to determine whether progress of the time lapse making process are printed
out to the console.}

\item{...}{Other parameters to be passed to plot_single_timeunit. Eg. highlight, alpha, size.}
}
\description{
Function to create a time lapse animation out of a data table of nodes.
}

